+++ tango/src/games/bn5.rs
@@ -11,7 +11,7 @@ impl games::Game for EXE5BImpl {
         (b"BRBJ", 0x00)
     }
 
-    fn family_and_variant(&self) -> (&str, u32) {
+    fn family_and_variant(&self) -> (&str, u8) {
         ("exe5", 0)
     }
 
@@ -50,7 +50,7 @@ impl games::Game for EXE5CImpl {
         (b"BRKJ", 0x00)
     }
 
-    fn family_and_variant(&self) -> (&str, u32) {
+    fn family_and_variant(&self) -> (&str, u8) {
         ("exe5", 1)
     }
 
@@ -89,7 +89,7 @@ impl games::Game for BN5PImpl {
         (b"BRBE", 0x00)
     }
 
-    fn family_and_variant(&self) -> (&str, u32) {
+    fn family_and_variant(&self) -> (&str, u8) {
         ("bn5", 0)
     }
 
@@ -128,7 +128,7 @@ impl games::Game for BN5CImpl {
         (b"BRKE", 0x00)
     }
 
-    fn family_and_variant(&self) -> (&str, u32) {
+    fn family_and_variant(&self) -> (&str, u8) {
         ("bn5", 1)
     }
